Analyzer refactoring
